se modifico mundo para que se pueda implemetar completamente el  enemigo para no afectar a los demas
