+++ src/Experience.jsx
@@ -2,7 +2,7 @@ import React from 'react'
 import { RoutesGame } from "./routes/RoutesGame";
 import { AvatarProvider } from './context/AvatarContext';
 import { AuthProvider } from './context/AuthContext';
-import { LifesProvider } from './context/ManagementLifes';
+import { LifesProvider } from './pages/level1/ManagementLifes';
 
 const Experience = () => {
   return (
