+++ src/pages/level1/Figures/Sphere.jsx
@@ -1,7 +1,7 @@
 import { RigidBody } from '@react-three/rapier';
 import React, { useRef } from 'react';
 import { useFrame } from '@react-three/fiber';
-import { useLifes } from '../../../context/ManagementLifes';
+import { useLifes } from '../../../pages/level1/ManagementLifes';
 
 export const Sphere = (props) => {
   const sphereBody = useRef();
