+++ src/pages/level1/Level1.jsx
@@ -13,7 +13,7 @@ import { Pane } from '../level1/layout/Pane';
 import { Coins } from '../level1/Figures/Coins';
 import useMovements from "../../utils/key-movements";
 import { Sphere } from './Figures/Sphere';
-import { useLifes } from '../../context/ManagementLifes';
+import { useLifes } from '../../pages/level1/ManagementLifes';
 import CharacterHudcueva_encantada from "./hud/CharacterHud";
 import GameOver from "./world/GameOver";
 import { Model } from './Figures/enemigo';
